DROID-2135 upload blocks
