+++ core-ui/src/main/java/com/anytypeio/anytype/core_ui/features/editor/holders/upload/MediaUpload.kt
@@ -1,29 +1,27 @@
 package com.anytypeio.anytype.core_ui.features.editor.holders.upload
 
 import android.view.View
-import androidx.core.view.updateLayoutParams
-import androidx.core.view.updatePadding
-import androidx.recyclerview.widget.RecyclerView
-import com.anytypeio.anytype.core_ui.BuildConfig
-import com.anytypeio.anytype.core_ui.R
+import android.widget.FrameLayout
+import com.anytypeio.anytype.core_ui.databinding.ItemBlockMediaUploadingBinding
 import com.anytypeio.anytype.core_ui.features.editor.BlockViewDiffUtil
 import com.anytypeio.anytype.core_ui.features.editor.BlockViewHolder
 import com.anytypeio.anytype.core_ui.features.editor.EditorTouchProcessor
 import com.anytypeio.anytype.core_ui.features.editor.SupportCustomTouchProcessor
-import com.anytypeio.anytype.core_utils.ext.dimen
+import com.anytypeio.anytype.core_ui.features.editor.decoration.DecoratableCardViewHolder
+import com.anytypeio.anytype.core_ui.features.editor.decoration.EditorDecorationContainer
+import com.anytypeio.anytype.core_ui.features.editor.decoration.applySelectorOffset
 import com.anytypeio.anytype.presentation.editor.editor.listener.ListenerType
 import com.anytypeio.anytype.presentation.editor.editor.model.BlockView
 
 abstract class MediaUpload(
-    view: View
-) : BlockViewHolder(view),
+    val binding: ItemBlockMediaUploadingBinding
+) : BlockViewHolder(binding.root),
     BlockViewHolder.IndentableHolder,
-    SupportCustomTouchProcessor {
+    SupportCustomTouchProcessor, DecoratableCardViewHolder {
 
-    abstract val root: View
+    override val decoratableContainer: EditorDecorationContainer get() = binding.decorationContainer
+    override val decoratableCard: View get() = binding.card
     abstract fun uploadClick(target: String, clicked: (ListenerType) -> Unit)
-    abstract override fun indentize(item: BlockView.Indentable)
-    abstract fun select(isSelected: Boolean)
 
     override val editorTouchProcessor = EditorTouchProcessor(
         fallback = { e -> itemView.onTouchEvent(e) }
@@ -33,11 +31,10 @@ abstract class MediaUpload(
         itemView.setOnTouchListener { v, e -> editorTouchProcessor.process(v, e) }
     }
 
-    fun bind(
+    open fun bind(
         item: BlockView.Upload,
         clicked: (ListenerType) -> Unit
     ) {
-        indentize(item)
         select(item.isSelected)
         with(itemView) {
             setOnClickListener { uploadClick(item.id, clicked) }
@@ -51,5 +48,22 @@ abstract class MediaUpload(
                 select(item.isSelected)
             }
         }
+    }
+
+    fun select(isSelected: Boolean) {
+        binding.selected.isSelected = isSelected
+    }
+
+    @Deprecated("Pre-nested-styling legacy.")
+    override fun indentize(item: BlockView.Indentable) {
+        // Do nothing.
+    }
+
+    override fun applyDecorations(decorations: List<BlockView.Decoration>) {
+        super.applyDecorations(decorations)
+        binding.selected.applySelectorOffset<FrameLayout.LayoutParams>(
+            content = binding.card,
+            res = itemView.resources
+        )
     }
 }
