+++ core-ui/src/main/java/com/anytypeio/anytype/core_ui/features/editor/holders/upload/PictureUpload.kt
@@ -1,38 +1,22 @@
 package com.anytypeio.anytype.core_ui.features.editor.holders.upload
 
-import android.view.View
-import com.anytypeio.anytype.core_ui.BuildConfig
 import com.anytypeio.anytype.core_ui.R
-import com.anytypeio.anytype.core_ui.databinding.ItemBlockPictureUploadingBinding
+import com.anytypeio.anytype.core_ui.databinding.ItemBlockMediaUploadingBinding
+import com.anytypeio.anytype.core_ui.extensions.drawable
 import com.anytypeio.anytype.core_ui.features.editor.decoration.DecoratableCardViewHolder
-import com.anytypeio.anytype.core_ui.features.editor.decoration.DecoratableViewHolder
-import com.anytypeio.anytype.core_ui.features.editor.decoration.EditorDecorationContainer
-import com.anytypeio.anytype.core_utils.ext.dimen
-import com.anytypeio.anytype.core_utils.ext.indentize
 import com.anytypeio.anytype.presentation.editor.editor.listener.ListenerType
 import com.anytypeio.anytype.presentation.editor.editor.model.BlockView
 
 class PictureUpload(
-    private val binding: ItemBlockPictureUploadingBinding
-) : MediaUpload(binding.root), DecoratableViewHolder, DecoratableCardViewHolder {
-
-    override val root: View = itemView
-
-    override val decoratableContainer: EditorDecorationContainer
-        get() = binding.decorationContainer
-
-    override val decoratableCard: View get() = binding.card
+    binding: ItemBlockMediaUploadingBinding
+) : MediaUpload(binding), DecoratableCardViewHolder {
 
     override fun uploadClick(target: String, clicked: (ListenerType) -> Unit) {
         clicked(ListenerType.Video.Upload(target))
     }
 
-    @Deprecated("Pre-nested-styling legacy.")
-    override fun indentize(item: BlockView.Indentable) {
-        // Do nothing.
-    }
-
-    override fun select(isSelected: Boolean) {
-        root.isSelected = isSelected
+    override fun bind(item: BlockView.Upload, clicked: (ListenerType) -> Unit) {
+        super.bind(item, clicked)
+        binding.fileIcon.setImageDrawable(itemView.context.drawable(R.drawable.ic_picture_inactive))
     }
 }
