+++ core-ui/src/main/java/com/anytypeio/anytype/core_ui/features/editor/holders/upload/VideoUpload.kt
@@ -1,47 +1,22 @@
 package com.anytypeio.anytype.core_ui.features.editor.holders.upload
 
-import android.view.View
-import android.widget.FrameLayout
-import com.anytypeio.anytype.core_ui.BuildConfig
 import com.anytypeio.anytype.core_ui.R
-import com.anytypeio.anytype.core_ui.databinding.ItemBlockVideoUploadingBinding
+import com.anytypeio.anytype.core_ui.databinding.ItemBlockMediaUploadingBinding
+import com.anytypeio.anytype.core_ui.extensions.drawable
 import com.anytypeio.anytype.core_ui.features.editor.decoration.DecoratableCardViewHolder
-import com.anytypeio.anytype.core_ui.features.editor.decoration.EditorDecorationContainer
-import com.anytypeio.anytype.core_ui.features.editor.decoration.applySelectorOffset
-import com.anytypeio.anytype.core_utils.ext.dimen
-import com.anytypeio.anytype.core_utils.ext.indentize
 import com.anytypeio.anytype.presentation.editor.editor.listener.ListenerType
 import com.anytypeio.anytype.presentation.editor.editor.model.BlockView
 
 class VideoUpload(
-    val binding: ItemBlockVideoUploadingBinding
-) : MediaUpload(binding.root), DecoratableCardViewHolder {
-
-    override val root: View = itemView
-
-    override val decoratableContainer: EditorDecorationContainer
-        get() = binding.decorationContainer
-
-    override val decoratableCard: View = binding.card
+    binding: ItemBlockMediaUploadingBinding
+) : MediaUpload(binding), DecoratableCardViewHolder {
 
     override fun uploadClick(target: String, clicked: (ListenerType) -> Unit) {
         clicked(ListenerType.Video.Upload(target))
     }
 
-    @Deprecated("Pre-nested-styling legacy.")
-    override fun indentize(item: BlockView.Indentable) {
-        // Do nothing.
-    }
-
-    override fun select(isSelected: Boolean) {
-        binding.selected.isSelected = isSelected
-    }
-
-    override fun applyDecorations(decorations: List<BlockView.Decoration>) {
-        super.applyDecorations(decorations)
-        binding.selected.applySelectorOffset<FrameLayout.LayoutParams>(
-            content = binding.card,
-            res = itemView.resources
-        )
+    override fun bind(item: BlockView.Upload, clicked: (ListenerType) -> Unit) {
+        super.bind(item, clicked)
+        binding.fileIcon.setImageDrawable(itemView.context.drawable(R.drawable.ic_video_inactive))
     }
 }
